setupcfg/pyprojtoml and factored into layers/models folder
