+++ .gitignore
@@ -1,2 +1,4 @@
 __pycache__/*
 .ipynb_checkpoints/*
+*.egg-info/*
+*/__pycache__/*
