+++ setup.cfg
@@ -1,0 +1,11 @@
+[metadata]
+name = mimicnet
+version = 0.0.1
+
+[options]
+packages = find:
+
+install_requires =
+    numpy
+    networkx
+    matplotlib
